requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
pydantic==2.11.7
paho-mqtt==2.1.0
python-dotenv==1.1.1
